(#4914) Improved stubbing in mount/parsed_spec tests.
A few of the spec tests were attempting to stub
Puppet::Type::Mount#default_target so that it pointed to a temporary
file rather than /etc/fstab, but they were creating the stub after the
first call to default_target, so both /etc/fstab and the temporary
file were being read.  This caused errors when running spec tests on
platforms where /etc/fstab is unreadable by non-privileged users.

Fixed the problem by moving the stub declaration earlier in the test.