Provider of Python packages via `pip`.
The details and history of this file are available on GitHub at <https://github.com/rcrowley/puppet-pip/blob/master/lib/puppet/provider/package/pip.rb>.  It has been available as a standalone gem.  I think this provider should become a part of the standard collection of providers now that it's clear the Python community are united behind it as their package manager.