(#6582) add fixture helper methods to replace unit test code.
We validate that fixtures exist, when requested, or that they match something
when they use a glob.  This catches internal errors where we don't match any
fixtures with a glob; this can reveal problems that would otherwise avoid all
assertions and result in internal failures.

The code is hidden out in a module, included in the main RSpec namespace.
This doesn't clean up the API any, but it isolates the code more effectively
and keeps the configuration file cleaner.

Reviewed-By: Nick Lewis <nick@puppetlabs.com>