(#6582) Eliminate the last vestige of the unit tests from the specs.
This eliminates a stub module in spec_helper, which was used for compatibility
with code that didn't bother to require the right files out of the unit tests.

It also removes test/lib from LOAD_PATH when running specs, since we don't run
with that code, and now we detect that as a larger scale test failure.

Reviewed-By: Nick Lewis <nick@puppetlabs.com>