(#6582) unstub Puppet settings to improve indirector queue tests.
This code was stubbing Puppet settings, which is no longer required now we
reset them between tests.  Using the real thing reduces points that the rest
of the code can break, too.

As a side effect this caused Puppet[:trace] to be "true", which meant that we
emitted a huge, nasty backtrace during the testing of a specific internal
failure handling case.

Reviewed-By: Nick Lewis <nick@puppetlabs.com>