(#6582) Don't demand the checkout be named 'puppet'.
As part of implementing the fixture support I hard-coded the assumption that
the git checkout was a directory named 'puppet'; this broke on our CI server,
and would break for anyone else who didn't follow that default.

This commit eliminates that assumption and depends only on the appropriate
part of the input filename.

Reviewed-By: Paul Berry <paul@puppetlabs.com>