maint: Fix a randomization test failure
The commit for #2597 included a test that asserted the text resulting
from detecting a cycle.  However, the cycle detection could start
randomly from any node, resulting in different text in the error.  I'm
not sure what the randomization key would be based on since the test
failed consistently over dozens of runs for me, and didn't for Daniel.

Paired-with: Daniel Pittman <daniel@puppetlabs.com>