maint: spec/integration/configurer has races in time checking.
This code depended on the time written by the configurer, and Time.now about
half a dozen lines later, being the same integer value.  This would work
almost all the time, because our code is pretty fast, but can race.

We now capture the time on both sides of the call, and verify that the emitted
time was in that range; this actually tests the whole purpose much better, and
more thoroughly.

Reviewed-by: Paul Berry <paul@puppetlabs.com>