(#6641) fix mount provider tests broken in the 2.6 merge.
A number of fixture-related changes were made in next, but not 2.6, which
resulted in extra tests added during 2.6 being broken.  The main thrust of
this change is to fix those tests by porting the fixture support code to the
newer helpers.

This also highlighted some shortfalls in our platform fixtures, which we
extend to add a bunch of additional fixtures, and to uniformly walk those to
perform fstab vs mount testing.