maint: Remove spec run noise
There were some warnings and stack traces in the spec output that aren't
necessary.

The only interesting fix is of the message:

lib/puppet/module.rb:79 warning: multiple values for a block parameter (0 for 1)
   from lib/puppet/util/logging.rb:30

If you call any form of logging on a module you end calling the file
method on the module just because logging always checks for that method
and calls it if it's defined, but in this case it's not defined in the
way that logging expected so passes the wrong paramters.

The easy solution is just to call logging on Puppet, which makes sense
in this case anyway, and I don't think it's worth a separate ticket to
deal with that logging warning.

Reviewed-by: Nick Lewis <nick@puppetlabs.com>