(#6641) Make it easier to add future platforms to the suite.
When a new fstab fixture is added, we also want to try to test mount output.
If that fixture is missing, we would have failed a test, but now mark it
pending.

This is more correct when, for example, we don't yet have that fixture: it
isn't technically a failure, just an incomplete set of test data.