Fixed #2645 - Added support for creating system users
On Red Hat, Ubuntu, Debian and deriatives the -r flag
allows creation of "system" users with a UID below that
defined in /etc/login.defs.

This commit adds support for a system parameter and a
system_users feature which can be used like so:

    user { "foo":
      system => true,
      ensure => present,
    }

This will create a user with a lower UID.

The system parameter defaults to false.