(#4884) Fix Test::Unit exec tests
Mostly this is whitespace cleanup, but other than that it's changing a
couple method names and calling run on the new exec providers instead of
the types.

We started moving these tests into spec, but they weren't very self
contained so were hard to map over cleanly.  For now leaving them since
they serve as a more integration level set of tests.

Paired-with: Max Martin