(#6722) load all functions before testing...
We historically had a state-dependency across tests in the parser function
support area; the first test caused the function to be loaded, and other tests
would then fail as a consequence of that.

We now autoload all functions at the top of each test suite, allowing us to
correctly and sensibly test on them as we should.

This theoretically prevents us from testing the autoloader implicitly in these
tests, but that should be tested independently.

Paired-With: Nick Lewis <nick@puppetlabs.com>