(#6658) Propagate ENC connection errors to the agent
When the master failed to run the External Node Classifier script it
would log an error on the master, but the agent didn't get back the full
error or the stack trace for it's logs.  By raising when there's an ENC
script problem on the master, this causes the error messages to
propagate to the agent.

Paired-with: Jacob Helwig <jacob@puppetlabs.com>