(#4576) Raise an error when a node is classified into a non-existent class
In 2.6.x, this was upgraded from "info" to "warning". This change for
Statler escalates the warning to an exception which will abort the
compile. This makes compiling fail consistently when you try to use an
undefined class from any of: node classifiers, the class keyword, and the
include function.

Paired-with: Jacob Helwig <jacob@puppetlabs.com>