MAINT: fix the misordered invocations in action.
When initializing we need to set the name and interface before we do anything
else, since the reasonable assumption for users is that those invariants are
there when their setter is called.

This allows someone to override the interface or name by misusing the call to
new, but they could already screw up by passing the wrong values, so whatever.