(#6749) implementing option handling in CLI string wrapper
The purpose of this is to adapt the generic option support in our strings to
the command line; we adapt the generic option information to optparse, and
establish our environment early in the process to ensure that we can play nice
with Puppet::Application for the moment.

In the process we ensure that we detect, and report, conflicts in option
naming across the board.  Additionally, when an option is declared with
multiple aliases, we insist that either all, or none, of them take an
argument.

To support this we support introspecting options having an optional argument,
as well as documentation and all.

Reviewed-By: Pieter van de Bruggen <pieter@puppetlabs.com>