(6911) Use normal methods to implement "depthfirst?" test
There was a class instance variable that was used to determine if a resource
types's children should be processed before or after the parent, to support
the one type (tidy) which did this.  Instead, we define a normal function in
Type to return false and override it in Tidy to return true.

Paired-with: Jesse Wolfe