(#6749) disable Action#invoke for this release.
So, we have Action#invoke, but it binds to the declaring class, not to the
correct instance.  Solving all the subtle issues around threads, global state,
and bindings without causing us too much pain is actually pretty hard, so
instead we pull the feature.

It can be enabled again in a future release and, being a strict extension
feature, we can do that without overly hurting anyone.  We still have full
access to the invocation through a marginally less pleasant syntax, but one
that people MUST be able to arrange to allow invoke to work, so on that front
we are clean.

Reviewed-By: Pieter van de Bruggen <pieter@puppetlabs.com>