(#6749) base indirector string should fail on invalid terminus.
We used to generate an info-level message, then carry on, which typically
resulted in raising because 'nil' doesn't implement the expected method that
we immediately try to call.

So, instead, raise a clear error at the time we fail to load, which gives a
pleasant rather than confusing error to the user.  Which at least means they
know why things have gone wrong...

Reviewed-By: Pieter van de Bruggen <pieter@puppetlabs.com>