(#5437) Invalidate cached TypeCollection when there was an error parsing
The caching of the TypeCollection in Puppet::Node::Environment would cause
parse errors to occur (and be reported) only once and never again, until
the file had changed on disk.  This would also cause empty catalogs to be
sent down to the agents further hiding the problem.

Now, when a file fails to parse, it will be re-parsed every time on every
following compilation, causing the parse error to be reported every time,
and preventing sending down empty catalogs to agents.

Paired-with: Nick Lewis <nick@puppetlabs.com>