(#6760) set terminus in indirector string base class.
We now accept a terminus option to each invocation, and set the terminus based
on that call.  This is probably incomplete, because it only sets the terminus
when given, and doesn't try to reset it to the default afterwards.

This also resets the terminus class after every invocation, to stop it leaking
state across calls.  This make, sadly, have some effects if you are not just
using the strings to invoke the terminus, but it beats having the strings
broken as well...

Reviewed-By: Pieter van de Bruggen <pieter@puppetlabs.com>