(6911) Add bookkeeping facade around Transaction#relationship_graph
To implement graph frontiers transactions need to track information about the
catalog's relationship graph.  For various reasons (serialzation, lifetime,
etc.) the data belongs with the transaction rather than the catalog or its
relationship graph.  This commit introduces a facade around the property used
to cheat Demeter which has the apropriate lifetime and can be used to hold the
state information durring a traversal.

Paired-with: Jesse Wolfe