(#6995) more robust testing of string loading
We were flushing loaded string versions in the test suite, but were not
clearing out the $" array used to determine if we had previously loaded that
file.

This lead to situations where tests cross-polluted because require though the
code was in memory, but the effects of loading that code had all been rolled
back.

Reviewed-By: Dan Bode <dan@puppetlabs.com>