maint: mocking 'require' causes random stack-overflow failures.
It turns out that mocking require on objects can cause random "stack depth
exceeded" failures, usually triggered by otherwise innocent changes to load
order or the use of require in the rest of the code.

Disable the currently failing test, which needs rewriting, and let it go for
now.  We need to systematically purge these problems once the build is stable.

Paired-With: Dan Bode <dan@puppetlabs.com>