maint: Add Array combinations method
Ruby < 1.8.7 doesn't have this method and we're using it in a test, so
tests won't run on 1.8.6 until this is in place.

It's probably a good thing to use much in implementation since it's
written in pure Ruby when using < 1.8.7 and in C when in > 1.8.7, but
then if you're using older Rubies you're probably not expecting much for
performance anyway.

Reviewed-by: Daniel Pittman <daniel@puppetlabs.com>