maint: disable garbage collector during individual test cases.
This reduced overhead of garbage collection from 50 percent to 20 percent in
my test runs, and reduced wall-time to match.  This seems to be a reasonable
win: we allow the GC to run on demand, but only outside the testing.

Memory use went from ~ 300MB to ~550MB between runs, which suggests that we
are generating a *lot* of garbage at times, but that we also benefit from
cleaning it automatically on a regular basis.

Reviewed-By: Matt Robinson <matt@puppetlabs.com>