(#6992) Expose available_subcommands as a class method.
We previously treated the list of legacy commands as an instance property of
Puppet::Util::CommandLine.  This made it difficult to obtain that information
outside the context of the instantiated application.

In the longer term this should wither and die as we eliminate the legacy
applications entirely, but until then exposing that on the class is the
lightest mechanism for making that useful.

Paired-With: Matt Robinson <matt@puppetlabs.com>