(#6962) delegate global usage to the help face.
The global --help handler (also invoked when puppet is run without any other
command line options at all) used to spit out a brief and generally not so
helpful message.

Now that we have a help face that can provide the same information in a much
more user-friendly form, we should delegate the function to that when
required.

Reviewed-By: Matt Robinson <matt@puppetlabs.com>