(#6962) Move option handling into #parse_options, not #preinit.
Logically, the extra work around option parsing for faces belongs in the
application parse_options method, not hidden in the step before.  This commit
moves that to the right place and fixes the fallout from that strange early
design decision.

Along the way we unify error reporting for invalid options so that all the
code paths result in the same externally detected failures.

Reviewed-By: Matt Robinson <matt@puppetlabs.com>