maint: ensure we handle '-foo=' options correctly in faces.
This is legal, and defines an option with an argument for optparse; we want to
make sure that we stay consistent with that in our code handling the same sort
of inputs.

This also adds a framework for later testing any other strange naming bits and
bobs we find around option handling.

Reviewed-By: Pieter van de Bruggen <pieter@puppetlabs.com>