(#7059) handle inherited action binding scope
We implemented the naive version of inheritance in the faces, in which an
action declared on a superclass remains bound to that even when accessed
from a subclass.

As a consequence our visibility of options on inherited classes is actually
much more limited than it should be, and the actions we inherit would never
see the options they should correctly have.

To fix this we correct the binding bug and handle lookup correctly over
inheritance in our code.

Reviewed-By: Pieter van de Bruggen <pieter@puppetlabs.com>