(#4258) Fix pkgutil spec test to have the correct provider
The resource being used for testing didn't explicitly set the provider,
so it ended up using whatever the default provider was on the system on
which it was run.  This was problematic when running the specs on a Mac
since the default provider is pkgdmg and that provider doesn't seem to
be upgradeable.  So when you tried:

    @resource[:ensure] = :latest

You got the error:

    1) Puppet::Type::Package::ProviderPkgutil when installing should use a command without versioned package
     Failure/Error: @resource[:ensure] = :latest
     Puppet::Error: Parameter ensure failed: Provider must have features 'upgradeable' to set 'ensure' to 'latest'
     # ./lib/puppet/parameter.rb:171:in `fail'     # ./lib/puppet/parameter.rb:257:in `validate'
     # ./lib/puppet/property.rb:300:in `should='     # ./lib/puppet/property.rb:300:in `each'
     # ./lib/puppet/property.rb:300:in `should='     # ./lib/puppet/property.rb:337:in `value='
     # ./lib/puppet/type.rb:416:in `[]='     # ./spec/unit/provider/package/pkgutil_spec.rb:35

Reviewed-by: Daniel Pittman <daniel@puppetlabs.com>