maint: speed up testing output of the help face.
We were generating the help output multiple times and testing it for multiple
properties; now we generate it once and test it multiple times.  This makes
for less clear error reporting, but saving dozens of calls at ~ 1/3rd a second
each is worth it.

Paired-With: Max Martin <max@puppetlabs.com>