(#7013) better default rendering support for faces
We have some specific requirements around rendering, including the ability of
authors of actions to add nice, custom rendering.  To support that we want
solid "basic" rendering for human-focused output.

This implements that generic rendering correctly and to spec, to give a sound
basis that we can build on for extensible rendering.
(#7013) better default rendering support for faces

We have some specific requirements around rendering, including the ability of
authors of actions to add nice, custom rendering.  To support that we want
solid "basic" rendering for human-focused output.

This implements that generic rendering correctly and to spec, to give a sound
basis that we can build on for extensible rendering.

Reviewed-By: Max Martin <max@puppetlabs.com>