(#7013) Handle rendering modes out in the application layer.
We no longer establish the rendering mode in the actions; they just default to
"nothing", and let that flow on out to the application layer.  That lets the
facade we put before the face determine the default behaviour.

This is mostly a no-op down in the CLI side, but it makes it much easier to
integrate into MCollective, HTTP-API, and for other non-CLI users of Faces.

Reviewed-By: Max Martin <max@puppetlabs.com>