Fix #7084 Make the log messages produced by whits less confusing
This patch changes the logging behavior of Whits so that instead of
talking about whits directly, we refer to the Class, Stage, or recursive
resource that they are sentinals for.
In the case where a "completion" Whit is notified by a resource, getting
a notification at all is counterinutitive, so I've changed the output to
a "debug"-priority message that describes what's happening.

Reviewed-By: Nick Lewis <nick@puppetlabs.com>